Prova a mantenere classe Cell mettendo data membri int,int
Mi sono bloccato all'introduzione della coordinata del Robot per mantenere la classe Cell definita però con due interi come dati membri... Non so come ottenere la coordinata del robot a partire dalla cella quando calcola la distanza tra 2 elementi come cella attuale-ostacolo o cella attuale-goal.

+++ Cell.h
@@ -12,7 +12,7 @@ class Cell{
 
 public:
     //costruttore e costruttore di default
-    Cell(double xC, double yC);
+    Cell(int xC, int yC);
     Cell();
 
     //funzioni membro, che non modificano l'oggetto
@@ -26,7 +26,7 @@ public:
     Cell path_planning_robot(double _eta, double _zeta, double _max_dist_infl, double dimGrid, const Cell& cgoal, const vector<Obstacle>& vecobst_pp) const;
 
 private:
-    double xC, yC;
+    int xC, yC;
 
 
 };
